% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomps.R
\name{cor_dist}
\alias{cor_dist}
\title{Compute pairwise correlation-based distance}
\usage{
cor_dist(X, method = "pearson", scale = TRUE, base = 1e+06,
  log_trans = TRUE, log_base = 10)
}
\arguments{
\item{X}{A data matrix, e.g. gene expression}

\item{method}{a character string indicating which correlation coefficient
is to be computed. One of "pearson" (default), "kendall", or
"spearman": can be abbreviated.}

\item{scale}{A boolean indicating whether to normalize the
columns (samples) of the data to the even sum.}

\item{base}{A numeric value for the shared column sum, if scale is TRUE.}

\item{log_trans}{A boolean indicating whether to log transform
the data prio to distance computation (log(X + 1)). Default is FALSE.}

\item{log_base}{A number indicating base for log transformation.
Default is 10.}
}
\value{
A dissimilarity matrix, D.
}
\description{
Pearson correlation distance for continuous
data: d = (1- cor(x_i, x_j))/2 where cor(x_i, x_j)
is the correlation between vector x_i and vector x_j.
}
