% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmodel.R
\name{fit_buds}
\alias{fit_buds}
\title{Fit Bayesian Unidimensional Scaling model to}
\usage{
fit_buds(D, K = NULL, method = c("vb", "mcmc"),
  hyperparams = list(tau_shape1 = 1, tau_shape2 = 1, gamma_epsilon = 2.5,
  gamma_bias = 2.5, gamma_rho_sq = 2.5, min_sigma = 0.001),
  init_from = c("random", "principal_curve"), seed = 1234,
  max_trials = 20, ...)
}
\arguments{
\item{D}{A square matrix of pairwise dissimilarities}

\item{K}{An integer indicating k-nearest neighbours to compute average
distances to kNN then used as weights in the power likelihood.}

\item{method}{Stan mode to use either "mcmc" or "vb"}

\item{hyperparams}{A list of BUDS hiperparameters with named components
set by default as follows "tau_shape1" = 1, "tau_shape2"= 1,
"gamma_epsilon" = 2.5, "gamma_bias" = 2.5, "gamma_rho_sq" = 2.5,
"min_sigma" = 0.001}

\item{init_from}{Initialization option either "random" or from a location
along "principal_curve" fitted to 2D representation using PCoA on D}

\item{seed}{A integer representing a random seed}

\item{max_trials}{A integer for maximum number of trials to fit data}

\item{...}{Other parameters for vb and stan}
}
\value{
The fitted `stanfit` model object
}
\description{
Fitting BUDS model to recover latent coordinates of data points from
observed/precomputed pairwise dissimilarities
}
