% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomps.R
\name{generic_dist}
\alias{generic_dist}
\title{Compute pairwise generic distances}
\usage{
generic_dist(X, method = c("manhattan", "euclidean", "exp manhattan",
  "exp euclidean", "maximum", "canberra", "binary", "minkowski",
  "jaccard"), log_trans = TRUE, log_base = 10, min_row_sum = 100,
  min_row_prevalence = 5)
}
\arguments{
\item{X}{A data matrix, e.g. gene expression}

\item{method}{a character string indicating which distance method to use.}

\item{log_trans}{A boolean indicating whether to log transform
the data prio to distance computation (log(X + 1)). Default is FALSE.}

\item{log_base}{A number indicating base for log transformation.
Default is 10.}

\item{min_row_sum}{A scalar indicating the minimum feature (species)
summ across columns (samples)  to be kept for distance computation.
Only used for Jaccard distance.}

\item{min_row_prevalence}{An integer indicating the minimum
prevalence (non-zero occurance) of a feature (species) across
samples to be kept for distance computation. Only used for Jaccard distance.}
}
\value{
A dissimilarity matrix, D.
}
\description{
A wrapper for dist function.
}
