% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_to_plot.R
\name{get_D_copies}
\alias{get_D_copies}
\title{Posterior samples of noisy dissimilarities}
\usage{
get_D_copies(D, buds_fit, B, min_sigma = 0.03)
}
\arguments{
\item{D}{A pairwise dissimilarity matrix.}

\item{buds_fit}{A stanfit object from fit_buds() output.}

\item{B}{An integer for the number of copies of D to
generate.}

\item{min_sigma}{A numeric parameter for minimum standard deviation
for the distance.}
}
\value{
A list of two list: one with posterior draws of noisy
dissimilarities, booD, and one of corresponding posterior
tau estimates bootTau.
}
\description{
Generate copies of noisy dissimilarity matrices
according to the estimates of latent coordinates
and the BUDS model.
}
