% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmodel.R
\name{get_dist_df}
\alias{get_dist_df}
\title{Reshape distance matrix to long format}
\usage{
get_dist_df(D, K = NULL)
}
\arguments{
\item{D}{A square matrix of pairwise dissimilarities}

\item{K}{An integer indicating k-nearest neighbours for computing average 
distance to kNN.}
}
\value{
A data frame to be used as input data for BUDS.
}
\description{
Internal function for getting distances in a long data.frame format,
a format used in fit_buds function
}
