% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_to_plot.R
\name{get_input_for_distatis}
\alias{get_input_for_distatis}
\title{Get data in a format for DiSTATIS}
\usage{
get_input_for_distatis(D, D.lst, tau_mode = NA, tau.lst = NULL,
  sample_data = NA)
}
\arguments{
\item{D}{A pairwise dissimilarity matrix.}

\item{D.lst}{A list of posterior noisy dissimilarity draws.}

\item{tau_mode}{[Optional] A mode estimate of tau.}

\item{tau.lst}{[Optional] A list of tau posterior draws
correspodning to each element of D.lst.}

\item{sample_data}{[Optional] A data frame with sample data}
}
\value{
A list with a 3D dissimilarity array, bootD,
and a list of data frames booData for each
slice of bootD and a data frame modeData
}
\description{
Get data in a format for DiSTATIS
}
