% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_to_plot.R
\name{get_subset_tau}
\alias{get_subset_tau}
\title{Data subset along gradient}
\usage{
get_subset_tau(tau, nkeep)
}
\arguments{
\item{tau}{A numeric vector representing the latent
positions of data points.}

\item{nkeep}{An integer indicating the number of samples
to keep.}
}
\value{
A vector with names of selected samples.
}
\description{
Select a subset of samples evenly spaced along
the underlying gradient using the latent coordinates
for the data points on a [0, 1] interval.
}
