% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_to_plot.R
\name{get_tau_df}
\alias{get_tau_df}
\title{Extract tau}
\usage{
get_tau_df(budsParams, prob = 0.95)
}
\arguments{
\item{budsParams}{A list of parameters extracted from BUDS stanfit object.}

\item{prob}{[Optional] A fraction for the probability of HPD interval.
Default is 0.95.}
}
\value{
A data frame with columns corresponding
to mode tau estimates and the lower and upper HPD.
}
\description{
Extracts tau parameter from the BUDS fit object. Computes
the mode and upper and lower highes posterior density
interval.
}
