% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmodel.R
\name{kNN_dist}
\alias{kNN_dist}
\title{k-nearest neighbours distance.}
\usage{
kNN_dist(D, K = NULL)
}
\arguments{
\item{D}{A dissimilarity matrix.}

\item{K}{An integer indicating the number of points 
to include in the neighbourhood. If not specified 
one tenth of the total number of data points is used.}
}
\value{
A list with mean distances to k-nearest 
neighbours and the associated variance of these
k-distances for each data point.
}
\description{
Compute average distance to k-nearest neighbours
for each data point from a dissimilarity matrix.
This approximates the data densities around 
each data point.
}
