% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{kNN_idx}
\alias{kNN_idx}
\title{Indices for k-nearest neighbours.}
\usage{
kNN_idx(D, K)
}
\arguments{
\item{D}{A dissimilarity matrix.}

\item{K}{An integer indicating the number of points 
to include in the neighbourhood. If not specified 
one tenth of the total number of data points is used.}
}
\value{
A data.frame with.
}
\description{
Function returns a n x K data.frame with every row
storing indices of the K nearest neighbour of 
a corresponding datapoint.
}
