% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_to_plot.R
\name{low_dim_vis}
\alias{low_dim_vis}
\title{Low dimensional data representation.
Compute reduced representations of the data.
A wrapper for cmdscale and Rtsne::Rtsne()}
\usage{
low_dim_vis(D, method = "PCoA", dims = 2, ...)
}
\arguments{
\item{D}{A pairwise dissimilarity matrix.}

\item{method}{[Optional] A dimensionality reduction
method either "PCoA" or "tSNE". Default is "PCoA".}

\item{dims}{[Optional] An integer for the number of dimensions
for to keep. Default is 2.}

\item{...}{other parameters for Rtsne::Rtsne() function}
}
\value{
A data frame with the low-dimensional
represenation.
}
\description{
Low dimensional data representation.
Compute reduced representations of the data.
A wrapper for cmdscale and Rtsne::Rtsne()
}
