% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{merge_heatmap_opts}
\alias{merge_heatmap_opts}
\title{Merge default options for a heatmap}
\usage{
merge_heatmap_opts(opts = list())
}
\arguments{
\item{opts}{[list] (Optional) A partially specified list used to customize
ppearance in ggplot theme(). Options that are already specified will not
be changed, those that are not will be filled in with defaults.}
}
\value{
opts [list]  A version of opts with unspecified options filled in
  with defaults.
}
\description{
Merge default options for a heatmap
}
