% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_binary_BUDS.R
\name{pet_binary_BUDS}
\alias{pet_binary_BUDS}
\title{Probability of Early Termination under a Two-Stage Simon BUDS}
\usage{
pet_binary_BUDS(n1, r1, p)
}
\arguments{
\item{n1}{Stage-1 sample size.}

\item{r1}{Stage-1 futility cutoff. The trial stops early for futility if
\code{r1} or fewer responses are observed during stage 1.}

\item{p}{Response probability where PET is evaluated.}
}
\value{
Numeric probability of early termination.
}
\description{
Computes the probability of early termination (stopping for futility
after stage 1) at a given response rate p.
}
