% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_buds_1D}
\alias{plot_buds_1D}
\title{Plot 1D latent coordinates}
\usage{
plot_buds_1D(tau_df, covariate = NULL, covariate_name = "covariate",
  color = "blue", color_label = NULL, idxBigger = NULL)
}
\arguments{
\item{tau_df}{A data frame with columns "tau", "tau_upper",
and "tau_lower".}

\item{covariate}{[Optional] If supplied taus coordiantes
are plotted against thissample  covariate vector.}

\item{covariate_name}{[Optional] The name of the covariat
supplied.}

\item{color}{[Optional] A vector or a character string for
point colors.}

\item{color_label}{[Optional] A character string for the color label.}

\item{idxBigger}{[Optional] A vector of integers indicating which
data point to highlight (make bigger).}
}
\value{
A ggplot2 object.
}
\description{
Plot 1D latent coordinates against their ranking
or supplied sample covariate data.
}
