% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_buds_trajectory}
\alias{plot_buds_trajectory}
\title{Plot data trajectory in 2 or3D}
\usage{
plot_buds_trajectory(budsParams, Y, eigs = NULL, sample_data = NULL,
  covariate_name = "covariate", path_col = "#2171B5", nPaths = 50,
  nCenters = 50)
}
\arguments{
\item{budsParams}{A list of parameters extracted from BUDS stanfit object.}

\item{Y}{A data frame with 2D representation of the data}

\item{eigs}{[Optional] A vector of eigenvalues corresponding
to columns of Y.}

\item{sample_data}{[Optional] A data frame with sample data.}

\item{covariate_name}{[Optional] A charcter string for the column
of sample_data to use for coloring points.}

\item{path_col}{[Optional] A charcter string for the color
of the posterior paths. Default is "#2171B5".}

\item{nPaths}{[Optional] An integer for number of
posterior paths to include. Default is 50.}

\item{nCenters}{[Optional] An integer for number of
data points to use for the highlighted "mode-path".
Default is 50.}
}
\value{
A ggplot2 object.
}
\description{
Plot 2 or 3D representations of the data with PCoA
of t-SNE, together with posterior data trajectories
(paths) computed with BUDS.
}
