% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_distatis}
\alias{plot_distatis}
\title{Plot DiSTATIS representation of the data}
\usage{
plot_distatis(distatis_df, consensus_df, color_label = NULL)
}
\arguments{
\item{distatis_df}{A data frame with partial configurations.}

\item{consensus_df}{A data frame with a consensus configuration.
consensus_df can contain a column "covariate" with sample data
that can be used for coloring.}

\item{color_label}{[Optional] A character string for the color label.}
}
\value{
A ggplot2 object.
}
\description{
Uses distatis_df data frame with configurations
computed by DiSTATIS for separate dissimilarity matrix
copies separetely, and a consensus_df data frame
for the "agreement" configuration to plot the data
representation with density clouds.
}
