% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_features_curves}
\alias{plot_features_curves}
\title{Plot features}
\usage{
plot_features_curves(X, tau, feat_idx = NA, log_trans = FALSE)
}
\arguments{
\item{X}{A data frame of matrix.}

\item{tau}{A vector with latent 1D coordinates for data points.}

\item{feat_idx}{An integer vector of selected features.}

\item{log_trans}{[Optional] A boolean for whether to log
transform the data before plotting. Default is FALSE.}
}
\value{
A ggplot2 object.
}
\description{
Plot feature trends along the estimated ordering
and include smoothing curves.
}
