% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_kNN_mean_dist}
\alias{plot_kNN_mean_dist}
\title{Plot distance to kNN}
\usage{
plot_kNN_mean_dist(D, tau, K = NULL, color = "blue",
  color_label = "Covariate", error_bars = FALSE)
}
\arguments{
\item{D}{A dissimilarity matrix.}

\item{tau}{A numeric vector of latent 1D coordinates.}

\item{K}{[Optional] An integer indicating the number of
points to include in the neighbourhood. If not specified
one tenth of the total number of data points is used.}

\item{color}{[Optional] A vector or a character string for
point colors.}

\item{color_label}{[Optional] A character string for the color label.}

\item{error_bars}{[Optional] Whether to include error bars in the plot.}
}
\value{
A ggplot2 object.
}
\description{
Computes average distance to K nearest neighbours and plots
the values for every data point ordered according to
coordinates computed with BUDS.
}
