% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot matrix}
\usage{
plot_matrix(mat, rowIdx = NULL, colIdx = NULL, ...)
}
\arguments{
\item{mat}{[matrix/ data.frame] (Required) Data matrix to plot.}

\item{rowIdx, colIdx}{[integer] (Optional) Default \code{NULL}.
Indices of the matrix to plot.}

\item{...}{(Optional) Additional plotting options.}
}
\value{
p [ggplot] A ggplot heatmap object with nice defaults.
}
\description{
A simple plot of matrix entries values.
}
\examples{
plot_matrix(matrix(rnorm(2000)*100, 100, 200))
}
