% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_ordered_matrix}
\alias{plot_ordered_matrix}
\title{Plot ordered data matrix}
\usage{
plot_ordered_matrix(X, tau, log_trans = FALSE, norm_samples = FALSE,
  keep_features = NULL, nfeatures = 500, byMean = TRUE, window = 5,
  minsum = 0)
}
\arguments{
\item{X}{A data frame of matrix where rows are features and columns
are samples.}

\item{tau}{A vector with latent 1D coordinates for data points.}

\item{log_trans}{[Optional] A boolean for whether to log
transform the data before plotting. Default is FALSE.}

\item{norm_samples}{A boolean whether the columns (samples) should be
normalize to even total sum.}

\item{keep_features}{[Optional] A list of features (rows) that must be
included in the heatmap.}

\item{nfeatures}{The number of features to keep. Default is 500.}

\item{byMean}{A boolean whether to order features by their "mean tau"
value. If FALSE the features are ordered by maximum in a window
along tau.}

\item{window}{An integer for the window size. Default is 5.}

\item{minsum}{A for the minimum sum for a feature in a data matrix.}
}
\value{
A ggplot2 object.
}
\description{
Plot a heatmap for a data matrix where columns
are reordered according to tau and rows according
to either byMean or by maximum value in a window
along tau.
}
