% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_reject_binary_BUDS.R
\name{pr_reject_binary_BUDS}
\alias{pr_reject_binary_BUDS}
\title{Exact Type I error for a Two-Stage Simon BUDS Design}
\usage{
pr_reject_binary_BUDS(n1, r1, n, r, p)
}
\arguments{
\item{n1}{Stage-1 sample size.}

\item{r1}{Stage-1 futility cutoff. The trial stops early for futility if
\code{r1} or fewer responses are observed during stage 1.}

\item{n}{Total sample size if the trial continues to the final analysis.}

\item{r}{Final futility cutoff. The treatment is not considered promising if
\code{r} or fewer responses are observed by the end of stage 2.}

\item{p}{Response probability where type I error is evaluated.}
}
\value{
Numeric type I error.
}
\description{
Computes the exact probability of rejecting the null hypothesis at a given
response rate p for a two-stage design.
}
