% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_to_plot.R, R/helper.R
\name{prin_curve_loc}
\alias{prin_curve_loc}
\title{Principal curve locations}
\usage{
prin_curve_loc(D, ...)

prin_curve_loc(D, ...)
}
\arguments{
\item{D}{A square matrix of pairwise dissimilarities}

\item{...}{Other parameters for princurve::principal_curve function.}

\item{\\dots}{Other parameters for princurve::principal.curve function.}

\item{D}{A square matrix of pairwise dissimilarities}
}
\value{
A numeric vector of positions along the fitted principal curve.

A numeric vector of positions along the fitted principal curve.
}
\description{
Estimating locations of data points projected on a principal curve fitted
to a PCoA 2D representation

Estimating locations of data points projected on a principal curve fitted
to a PCoA 2D representation
}
