% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_BUDS.R
\name{print.binary_BUDS}
\alias{print.binary_BUDS}
\title{Print Binary Simon BUDS Designs}
\usage{
\method{print}{binary_BUDS}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"binary_BUDS"}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Displays the input settings and selected binary two-stage Simon designs
stored in a \code{"binary_BUDS"} object.
}
