% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.binary_BUDS.R
\name{print.binary_BUDS_summary}
\alias{print.binary_BUDS_summary}
\title{Print a Summary of a Binary Simon BUDS Design}
\usage{
\method{print}{binary_BUDS_summary}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"binary_BUDS_summary"}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a Summary of a Binary Simon BUDS Design
}
