% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opchar_binary_BUDS.R
\name{print.opchar}
\alias{print.opchar}
\title{Print Operating Characteristics}
\usage{
\method{print}{opchar}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"opchar"}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Prints operating-characteristics results produced by
\code{\link[=opchar_binary_BUDS]{opchar_binary_BUDS()}} or \code{\link[=opchar_tte_BUDS]{opchar_tte_BUDS()}} for binary or time-to-event
two-stage designs.
}
