% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.opchar.R
\name{print.opchar_summary}
\alias{print.opchar_summary}
\title{Print a Summary of Operating Characteristics}
\usage{
\method{print}{opchar_summary}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"opchar_summary"}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a Summary of Operating Characteristics
}
