% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_BUDS.R
\name{print.tte_BUDS}
\alias{print.tte_BUDS}
\title{Print an r-KJ Design Object under BUDS}
\usage{
\method{print}{tte_BUDS}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"tte_BUDS"}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Displays the input settings and selected TTE designs stored in an
\code{"tte_BUDS"} object.
}
