% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tte_BUDS.R
\name{print.tte_BUDS_summary}
\alias{print.tte_BUDS_summary}
\title{Print a Summary of a Robust r-KJ Design under BUDS}
\usage{
\method{print}{tte_BUDS_summary}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"tte_BUDS_summary"}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a Summary of a Robust r-KJ Design under BUDS
}
