% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_to_plot.R
\name{run_distatis}
\alias{run_distatis}
\title{Distatis ordination coordiantes}
\usage{
run_distatis(bootD, dims = 2, booData.lst = NULL, modeData = NULL)
}
\arguments{
\item{bootD}{A 3D array with posterior dissimilarity samples.}

\item{dims}{An integer for the number of dimensions for
DiSTATIS.}

\item{booData.lst}{[Optional] A list of length equal dim(bootD)[3]
containing data frames corresponding to bootD slices.}

\item{modeData}{[Optional] A sample data frame for consensus
configuration.}
}
\value{
A list with two data frames: partial for DiSTATIS
configurations for each slice of bootD separately,
and consensus with an "average" configuration.
}
\description{
Compute ordination coordinates for all copies
of D using a three-way MDS method, DiSTATIS.
}
