% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.binary_BUDS.R
\name{summary.binary_BUDS}
\alias{summary.binary_BUDS}
\title{Summarize a Binary Simon Design under BUDS}
\usage{
\method{summary}{binary_BUDS}(object, design_name, ...)
}
\arguments{
\item{object}{Object of class \code{"binary_BUDS"}.}

\item{design_name}{Name of the design to summarize.}

\item{...}{Unused.}
}
\value{
An object of class \code{"binary_BUDS_summary"}.
}
\description{
Produces a prose summary for one selected design from a
\code{"binary_BUDS"} object.
}
