% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.opchar.R
\name{summary.opchar}
\alias{summary.opchar}
\title{Summarize Operating Characteristics}
\usage{
\method{summary}{opchar}(object, design_name = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{opchar}.}

\item{design_name}{Optional design name when \code{object} contains multiple
designs.}

\item{...}{Unused.}
}
\value{
An object of class \code{opchar_summary}.
}
\description{
Produces a prose summary from an \code{opchar} object returned by
\code{\link[=opchar_binary_BUDS]{opchar_binary_BUDS()}} or \code{\link[=opchar_tte_BUDS]{opchar_tte_BUDS()}}.
}
