% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tte_BUDS.R
\name{summary.tte_BUDS}
\alias{summary.tte_BUDS}
\title{Summarize a Survival r-KJ under BUDS}
\usage{
\method{summary}{tte_BUDS}(object, design_name, ...)
}
\arguments{
\item{object}{Object of class \code{"tte_BUDS"}.}

\item{design_name}{Name of the design to summarize.}

\item{...}{Unused.}
}
\value{
An object of class \code{"tte_BUDS_summary"}.
}
\description{
Produces a prose summary for one selected design from a
\code{"tte_BUDS"} object.
}
