% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomps.R
\name{transform_dist}
\alias{transform_dist}
\title{Rank based transform distances to triangular distribution}
\usage{
transform_dist(D, threshold = FALSE, increment = NULL)
}
\arguments{
\item{D}{A dissimilarity matrix.}

\item{threshold}{A boolean whether threshold ranking should be used.}

\item{increment}{A numeric scalar for increments for distance ranking.}
}
\value{
A transformed dissimilarity matrix, D.
}
\description{
Transforms the dissimularities to follow a triangular
distribution on [0, 1] interval to better match distances
on scalar coordinates that are approximately uniformly
distributed.
}
