% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buds-package.R
\docType{package}
\name{buds-package}
\alias{buds-package}
\alias{BUDS,}
\alias{buds}
\title{The 'buds' package.}
\description{
BUDS: package for Bayesian Unidimensional Scaling for mapping
high-dimensional objects to 1D latent coordinates using pairwise
dissimilarities.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
