% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buds-package.R
\docType{package}
\name{buds-package}
\alias{buds-package}
\alias{BUDS,}
\alias{buds}
\title{buds: package for Bayesian Unidimensional Scaling for mapping high-dimensional
objects to 1D latent coordinates using pairwise dissimilarities.}
\description{
buds: package for Bayesian Unidimensional Scaling for mapping high-dimensional
objects to 1D latent coordinates using pairwise dissimilarities.
}
